Remove false positives and clean up a bit

+++ Cerebrum/Group.py
@@ -28,8 +28,8 @@ that fashion.  Hence, this module **requires** the caller to supply a
 name when constructing a Group object."""
 
 
-import mx
 import six
+import mx
 
 import cereconf
 from Cerebrum import Utils
@@ -54,6 +54,25 @@ class Group(EntityQuarantine, EntityExternalId, EntityName,
         super(Group, self).clear()
         self.clear_class(Group)
         self.__updated = []
+
+    def exists(self, table=None, binds=None):
+        if not (table and binds):
+            raise ValueError("missing args")
+        where = 'AND '.join('{0}=:{0}'.format(x) for x in binds if x != 'g_id')
+        exists_stmt = """
+          SELECT EXISTS (
+            SELECT 1
+            FROM [:table schema=cerebrum name={table}]
+            WHERE {where}
+          )
+        """.format(table=table, where=where)
+        try:
+            self.query_1(exists_stmt, binds)
+            return True
+        except Errors.NotFoundError:
+            return False
+        except Errors.TooManyRowsError:
+            return True
 
     def populate(self, creator_id=None, visibility=None, name=None,
                  description=None, expire_date=None, parent=None):
@@ -87,6 +106,7 @@ class Group(EntityQuarantine, EntityExternalId, EntityName,
         self.write_db()
 
     def is_expired(self):
+        """Checks if group is expired"""
         now = mx.DateTime.now()
         if self.expire_date is None or self.expire_date >= now:
             return False
@@ -103,7 +123,7 @@ class Group(EntityQuarantine, EntityExternalId, EntityName,
     # lengths and max length in database is 256 characters
     def illegal_name(self, name, max_length=256):
         """Return a string with error message if groupname is illegal"""
-        if len(name) == 0:
+        if not name:
             return "Must specify group name"
         if len(name) > max_length:
             return "Name %s too long (%d char allowed)" % (name, max_length)
@@ -152,27 +172,29 @@ class Group(EntityQuarantine, EntityExternalId, EntityName,
                           'visib': int(self.visibility),
                           'creator_id': self.creator_id,
                           'exp_date': self.expire_date})
-            self._db.log_change(self.entity_id, self.clconst.group_create, None)
+            self._db.log_change(self.entity_id,
+                                self.clconst.group_create,
+                                None)
             self.add_entity_name(self.const.group_namespace, self.group_name)
         else:
-            cols = [('description', ':desc'),
-                    ('visibility', ':visib'),
-                    ('creator_id', ':creator_id'),
-                    ('expire_date', ':exp_date')]
-            self.execute("""
-            UPDATE [:table schema=cerebrum name=group_info]
-            SET %(defs)s
-            WHERE group_id=:g_id""" % {'defs': ", ".join(
-                ["%s=%s" % x for x in cols if x[0] != 'group_id'])},
-                {'g_id': self.entity_id,
-                 'desc': self.description,
-                 'visib': int(self.visibility),
-                 'creator_id': self.creator_id,
-                 'exp_date': self.expire_date})
-            self._db.log_change(self.entity_id,
-                                self.clconst.group_mod,
-                                None,
-                                change_params=self.__updated)
+            binds = {'desc': self.description,
+                     'group_id': self.entity_id,
+                     'visib': int(self.visibility),
+                     'creator_id': self.creator_id,
+                     'exp_date': self.expire_date}
+            if not self.exists(table='group_info', binds=binds):
+                # True positive
+                set_str = ', '.join(
+                    '{0}=:{0}'.format(x) for x in binds if x != 'group_id')
+                update_stmt = """
+                UPDATE [:table schema=cerebrum name=group_info]
+                SET {set_str}
+                WHERE group_id=:group_id""".format(set_str=set_str)
+                self.execute(update_stmt, binds)
+                self._db.log_change(self.entity_id,
+                                    self.clconst.group_mod,
+                                    None,
+                                    change_params=self.__updated)
             if 'group_name' in self.__updated:
                 self.update_entity_name(
                     self.const.group_namespace,
@@ -379,12 +401,16 @@ class Group(EntityQuarantine, EntityExternalId, EntityName,
             Group (id) to remove member from
         """
 
-        self.execute("""
-        DELETE FROM [:table schema=cerebrum name=group_member]
-        WHERE
-          group_id=:g_id AND
-          member_id=:m_id""", {'g_id': group_id,
-                               'm_id': member_id})
+        binds = {'group_id': group_id,
+                 'member_id': member_id}
+        if not self.exists(table='group_member', binds=binds):
+            # False positive
+            return
+        delete_stmt = """
+          DELETE FROM [:table schema=cerebrum name=group_member]
+          WHERE group_id=:group_id AND
+                member_id=:member_id"""
+        self.execute(delete_stmt, binds)
         self._db.log_change(group_id, self.clconst.group_rem, member_id)
 
     def search(self,
@@ -504,7 +530,7 @@ class Group(EntityQuarantine, EntityExternalId, EntityName,
             # workset contains ids of the entities that are members. in each
             # iteration we are looking for direct parents of whatever is in
             # workset.
-            workset = set([int(x) for x in member_id])
+            workset = set(int(x) for x in member_id)
             while workset:
                 tmp = workset
                 workset = set()
